+++ src/preload.js
@@ -4,9 +4,7 @@ import * as os from "os";
 
 const CONFIGS_FILE_PATH = os.homedir() + '/.watchdog/config.json'
 
-window.getTrayPosition = async () => {
-    return await ipcRenderer.invoke('get-tray-position')
-}
+window.appVersion = require('../package.json').version
 
 window.logger = {
     debug: (source, msg) => {
@@ -15,6 +13,10 @@ window.logger = {
     info: (source, msg) => {
         ipcRenderer.send('log-info', [source, msg])
     },
+}
+
+window.getTrayPosition = async () => {
+    return await ipcRenderer.invoke('get-tray-position')
 }
 
 window.log = (msg) => {
@@ -48,8 +50,8 @@ window.setAppIcon = (color) => {
     ipcRenderer.send('set-app-icon', color)
 }
 
-window.showNotification = (title, body) => {
-    ipcRenderer.send('show-notification', [title, body])
+window.showNotification = (subtitle, body) => {
+    ipcRenderer.send('show-notification', [subtitle, body])
 }
 
 window.quitApp = () => {