+++ CHANGELOG.md
@@ -1,1 +1,22 @@
-### First release
+# Release 1.0.1
+
+## New Functionality
+
+### Updates center
+When new version will be available you receive notification 
+and special icon appear in the top left corner.
+
+### Pause/Resume PRs fetching
+You can temporally pause/resume refreshing using button in the top right corner. 
+For example: on weekends to don't quit app at all.
+
+## Fixes
+Minor fixes for Windows platform
+
+Menu reloading on new data fetched
+
+PRs filtering
+
+## Other 
+
+### Optimized number of API requests